requirements.txt
annotated-doc==0.0.4
annotated-types==0.7.0
anyio==4.12.1
certifi==2026.2.25
charset-normalizer==3.4.5
click==8.1.8
exceptiongroup==1.3.1
filelock==3.19.1
fsspec==2025.10.0
h11==0.16.0
hf-xet==1.3.2
httpcore==1.0.9
httpx==0.28.1
huggingface_hub==0.36.2
idna==3.11
jsonpatch==1.33
jsonpointer==3.0.0
langchain-core==0.3.83
langchain-huggingface==0.3.1
langsmith==0.4.37
markdown-it-py==3.0.0
mdurl==0.1.2
orjson==3.11.5
packaging==25.0
pydantic==2.12.5
pydantic_core==2.41.5
Pygments==2.19.2
PyYAML==6.0.3
requests==2.32.5
requests-toolbelt==1.0.0
rich==14.3.3
shellingham==1.5.4
tenacity==9.1.2
tokenizers==0.22.2
tqdm==4.67.3
typer==0.23.2
typing-inspection==0.4.2
typing_extensions==4.15.0
urllib3==2.6.3
uuid_utils==0.14.1
zstandard==0.25.0
## The following requirements were added by pip freeze:
aiohappyeyeballs==2.6.1
aiohttp==3.13.3
aiosignal==1.4.0
altair==5.5.0
async-timeout==4.0.3
attrs==25.4.0
blinker==1.9.0
cachetools==6.2.6
dataclasses-json==0.6.7
distro==1.9.0
fastapi==0.128.8
frozenlist==1.8.0
gitdb==4.0.12
GitPython==3.1.46
groq==0.37.1
httpx-sse==0.4.3
Jinja2==3.1.6
joblib==1.5.3
jsonschema==4.25.1
jsonschema-specifications==2025.9.1
langchain==0.3.28
langchain-community==0.3.31
langchain-groq==0.3.8
langchain-text-splitters==0.3.11
MarkupSafe==3.0.3
marshmallow==3.26.2
mpmath==1.3.0
multidict==6.7.1
mypy_extensions==1.1.0
narwhals==2.18.0
networkx==3.2.1
numpy==2.0.2
pandas==2.3.3
pillow==11.3.0
propcache==0.4.1
protobuf==6.33.6
pyarrow==21.0.0
pydantic-settings==2.11.0
pydeck==0.9.1
pypdf==6.9.1
python-dateutil==2.9.0.post0
python-dotenv==1.2.1
pytz==2026.1.post1
referencing==0.36.2
regex==2026.1.15
rpds-py==0.27.1
safetensors==0.7.0
scikit-learn==1.6.1
scipy==1.13.1
sentence-transformers==5.1.2
six==1.17.0
smmap==5.0.3
sniffio==1.3.1
SQLAlchemy==2.0.48
starlette==0.49.3
streamlit==1.50.0
sympy==1.14.0
threadpoolctl==3.6.0
toml==0.10.2
torch==2.8.0
tornado==6.5.5
transformers==4.57.6
typing-inspect==0.9.0
tzdata==2025.3
yarl==1.22.0